Add TaskController
Add dto and dto mapper for Task

Add integration tests for TaskController

Add new method to ControllerAdvisor that handle TaskNotFoundException

Modify addTask method in ITaskService. Now this method takes only parameter.

Modify updateTask and deleteTask methods. Methods check who wants to modify or delete task.

+++ src/main/java/com/example/TaskManager/service/TaskService.java
@@ -1,10 +1,15 @@
 package com.example.TaskManager.service;
 
+import com.example.TaskManager.exception.ForbiddenException;
 import com.example.TaskManager.exception.TaskNotFoundException;
+import com.example.TaskManager.model.Role;
 import com.example.TaskManager.model.Task;
 import com.example.TaskManager.model.User;
 import com.example.TaskManager.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,9 +34,9 @@ public class TaskService implements ITaskService{
 
     @Override
     @Transactional
-    public Task addTask(Long userId, Task task) {
+    public Task addTask(Task task) {
 
-        User owner = userService.findUserById(userId);
+        User owner = userService.findUserById(task.getOwner().getId());
 
         task.setOwner(owner);
 
@@ -43,6 +48,15 @@ public class TaskService implements ITaskService{
     public Task updateTask(Long id, Task task) {
 
         Task taskToEdit = findTask(id);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+        User loggedUser = userService.findUserByUsername(userDetails.getUsername());
+
+        if(!taskToEdit.getOwner().getId().equals(loggedUser.getId())
+                && loggedUser.getRoles().stream().map(Role::getName).noneMatch(role -> role.equals("ADMIN")))
+            throw new ForbiddenException();
 
         if(task.getName() != null && !task.getName().equals(taskToEdit.getName()))
             taskToEdit.setName(task.getName());
@@ -62,6 +76,15 @@ public class TaskService implements ITaskService{
 
         if(!taskRepository.existsById(id))
             throw new TaskNotFoundException(id);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+
+        User loggedUser = userService.findUserByUsername(userDetails.getUsername());
+
+        if(loggedUser.getTasks().stream().map(Task::getId).noneMatch(userTask -> userTask.equals(id))
+                && loggedUser.getRoles().stream().map(Role::getName).noneMatch(role -> role.equals("ADMIN")))
+            throw new ForbiddenException();
 
         taskRepository.deleteById(id);
     }
